# ByteBufferLib, RecordBuffer: use bulk get/put APIs