@lucasvr 2023-03-16
@kinow not really; the function is short-lived and the amount of memory allocated hardly goes beyond a few kilobytes with the datasets I've tested.