@afs 2023-03-20
This is looking good. I tried the various `tdb2.tdbloader` choices. I got a ~10% improvement for loader=phased (the default from the command line) and loader=basic. test data was 25e6 BSBM triples. Less on loader=parallel but that is not surprising as it is not typically CPU limited.

Loading into Fuseki is, roughly, "loader=basic" to avoid an upload ruining the response for queries running at the same time. The bulk loaders like "phased" and "parallel" are standalone and don't support concurrent use.

I'll do some more testing with different data sizes and different warm-ups.

@lucasvr - for those timings, did you run a warm-up phase?