@afs 2023-03-20
Just to check on detail - it was a new server for each timing run?

Levels of optimization kick as the run happens, so warming up also causes the JIT to run. I think the optimizer spots patterns and can, for example, not do all the bound checking in a loop when ti knows the index can not be out of bounds.