@lucasvr 2023-03-20
The code profiler, which reported an excessive number of calls to `bbcopy2()`, sampled a single Fuseki process for several minutes as I ingested different TTL files via SPARQL. Only afterwards I benchmarked different instances of Fuseki by launching a new server for each run.

Does JIT depend on any particular version of JDK/JRE? Based on my observations it didn't seem to identify the hottest execution path and optimize `bbcopy2()`.