@afs 2023-03-21
Some optimization happens only after a large number of calls to a function. So there may be further optimizations that have not occurred.

The exact optimizations will be JVM dependent and get better with every release. (I've only used OpenJDK.)

The B+tree implementation was tested by generating random sequences and building tree then deleting it in another random odder, while checking internal consistency. Run 10's of millions of times while printing a dot every so many iterations of create/delete. The dots started appearing quite slowly, then a bit quicker then quicker again as various compile/optimize steps happened.