@afs 2023-03-27
@lucasvr - if you remove the `RecordBuffer` part, we can apply this PR.

Alternatively, I can take the PR diff and sort it out (but it does not have your name as the git-author).