@afs 2023-03-30
PR #1819 is the `ByteBufferLib` improvements from this PR.

I measued the sped improvement of #1819 and got a similar gain in performance.